+++ src/index.ts
@@ -3,6 +3,7 @@ import * as document from './api/document.ts';
 import * as graphics from './api/graphics.ts';
 import * as layout from './api/layout.ts';
 import * as makePdf from './api/make-pdf.ts';
+import * as pdfMaker from './api/PdfMaker.ts';
 import * as sizes from './api/sizes.ts';
 import * as text from './api/text.ts';
 
@@ -12,6 +13,7 @@ export const pdf = {
   ...graphics,
   ...layout,
   ...makePdf,
+  ...pdfMaker,
   ...sizes,
   ...text,
 };
@@ -21,5 +23,6 @@ export * from './api/document.ts';
 export * from './api/graphics.ts';
 export * from './api/layout.ts';
 export * from './api/make-pdf.ts';
+export * from './api/PdfMaker.ts';
 export * from './api/sizes.ts';
 export * from './api/text.ts';